% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_null_table.R
\name{load_zdip_null}
\alias{load_zdip_null}
\title{Load Z-Dip null table into the global environment if one so chooses}
\usage{
load_zdip_null(name = "dip_null_table")
}
\arguments{
\item{name}{Name to assign to the table in the user's environment}
}
\value{
Invisibly returns the null table.
}
\description{
Loads the precomputed null moments used by Z-Dip into the user's
global environment as `dip_null_table`.
}
