% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdip.R
\name{zdip}
\alias{zdip}
\title{Z-Dip statistic}
\usage{
zdip(
  x,
  downsample = FALSE,
  N0 = 1000,
  nsim = 200,
  squash = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{numeric vector}

\item{downsample}{logical; use downsampling for large samples}

\item{N0}{subsample size (if downsample = TRUE)}

\item{nsim}{number of subsamples (if downsample = TRUE)}

\item{squash}{logical; apply squashing function 2 / (1 + exp(-.595 * z)) - 1, to bound z in [-1, +1]}

\item{verbose}{logical; print informational messages}
}
\description{
Z-normalized version of Hartigan's Dip statistic.
}
